GRHB-203: * refactor: add FAB component
