+++ mobile/src/components/courses/components/components.ts
@@ -1,2 +1,1 @@
-export { AddCourseButton } from './add-course-button/add-course-button';
 export { CourseCard } from './course-card/course-card';
